save and include later in autocomplete original input value
After filling input with some value and selecting one of the entries
from autocomplete, we save previous/original value (before selecting
any of the entries) to local state to preserve for later use. We use
when are about to wrap around autocomplete entries, for example when
we're at last autocomplete entry and we press arrow down key. Instead
of focusing top entry, we populate input value with the original one
stored in local state. That way original entry is treated kinda like
another item in autocomplete.