fix: Use unique ids for each suggestion
Currently, on the mouseover event, each suggestion updates its id to be the active suggestion's id.
Each suggestion should have a unique id (this is a best practice and is also necessary for ARIA
compliance). This uses the specific suggestion's Google place id as part of the id so that each one
is unique.