Add test coverage for getLatLng
*Alter googlePayload to be closer to reality

The mocked payload being defined here adhered to a different interface
than what google actually returns.

I've changed the payload to be in line with what google returns.

I've made this change because I've seen other projects use this mocked
payload before and it causes confusion. I also plan on writing tests for
getLatLng which currently has a TODO for that task. Writing tests for
that function will require a more realisitic mocked payload.

https://developers.google.com/maps/documentation/javascript/reference/coordinates#LatLng

*Add test coverage for getLatLng function

I've added test coverage for the getLatLng function in this commit. This
will help ensure against regressions and provide developers using this
library with more confidence that it is throughly tested.